Make GCC 5.4 happy with those type casting
GCC 5.4 will report those type castings are not legal.

+++ src/codegen/irgen/hooks.cpp
@@ -351,7 +351,7 @@ CompiledFunction* compileFunction(FunctionMetadata* f, FunctionSpecialization* s
             break;
         }
         default:
-            RELEASE_ASSERT(0, "%d", effort);
+            RELEASE_ASSERT(0, "%d", static_cast<int>(effort));
     }
 
     // free the bjit code if this is not a OSR compilation
